requirements.txt
Django==4.1.1
python-dotenv==0.21.0
vk-api==11.9.9
Pillow==9.2.0
